Update number recognizer exports
