+++ JavaScript/recognizers-number/src/recognizers-text-number.ts
@@ -9,10 +9,15 @@ export { SpanishNumeric } from "./resources/spanishNumeric";
 export { Constants } from "./number/constants";
 export { IExtractor, ExtractResult, RegExpValue, BaseNumberExtractor, BasePercentageExtractor } from "./number/extractors";
 export { NumberMode, LongFormatType, AbstractNumberModel, NumberModel, OrdinalModel, PercentModel } from "./number/models";
-export { AgnosticNumberParserType, IParser, ParseResult, INumberParserConfiguration, BaseNumberParser, BasePercentageParser, AgnosticNumberParserFactory } from "./number/parsers";
+export { AgnosticNumberParserType, AgnosticNumberParserFactory } from "./number/agnosticNumberParser";
+export { IParser, ParseResult, INumberParserConfiguration, BaseNumberParser, BasePercentageParser } from "./number/parsers";
 export { EnglishCardinalExtractor, EnglishDoubleExtractor, EnglishFractionExtractor, EnglishIntegerExtractor, EnglishNumberExtractor, EnglishOrdinalExtractor, EnglishPercentageExtractor } from "./number/english/extractors";
 export { EnglishNumberParserConfiguration } from "./number/english/parserConfiguration";
 export { SpanishCardinalExtractor, SpanishDoubleExtractor, SpanishFractionExtractor, SpanishIntegerExtractor, SpanishNumberExtractor, SpanishOrdinalExtractor, SpanishPercentageExtractor } from "./number/spanish/extractors";
 export { SpanishNumberParserConfiguration } from "./number/spanish/parserConfiguration";
 export { PortugueseCardinalExtractor, PortugueseDoubleExtractor, PortugueseFractionExtractor, PortugueseIntegerExtractor, PortugueseNumberExtractor, PortugueseOrdinalExtractor, PortuguesePercentageExtractor } from "./number/portuguese/extractors";
 export { PortugueseNumberParserConfiguration } from "./number/portuguese/parserConfiguration";
+
+export { ChineseCardinalExtractor, ChineseDoubleExtractor, ChineseFractionExtractor, ChineseIntegerExtractor, ChineseNumberExtractor, ChineseOrdinalExtractor, ChinesePercentageExtractor } from "./number/chinese/extractors";
+export { ChineseNumberParserConfiguration } from "./number/chinese/parserConfiguration";
+export { ChineseNumberParser } from "./number/chinese/parsers";
